refactor: split core/lib/report into type/render/build/outcome layers
