+++ setup/docker/transparent/scripts/report-action.node.ts
@@ -9,10 +9,10 @@
  */
 import * as core from "@actions/core";
 import { createDocker } from "../../../../core/lib/docker/client.ts";
-import { buildReportParameters } from "../../../../core/lib/report/report-parameters.ts";
-import { buildTransparentReportData } from "../../../../core/lib/report/build-transparent-report-data.ts";
-import { renderReportMarkdown } from "../../../../core/lib/report/render-report-markdown.ts";
-import { emitBlockedOutcome } from "../../../../core/lib/report/emit-blocked-outcome.ts";
+import { buildReportParameters } from "../../../../core/lib/report/parameters.ts";
+import { buildTransparentReportData } from "../../../../core/lib/report/build/transparent.ts";
+import { renderReportMarkdown } from "../../../../core/lib/report/render/render-report-markdown.ts";
+import { emitBlockedOutcome } from "../../../../core/lib/report/outcome/emit-blocked-outcome.ts";
 import { errorMessage } from "../../../../core/lib/general/error-message.ts";
 import { writeStepSummary } from "../../lib/write-step-summary.ts";
